requirements.txt
# Time's Pixel - Required Dependencies

# Core visualization and graphics
pygame>=2.6.0
numpy>=1.21.0
pillow>=10.0.0

# Data processing and analysis
pandas>=1.3.0

# Animation and image processing
imageio>=2.19.0

# Additional utilities (optional but recommended)
matplotlib>=3.5.0
requests>=2.25.0